all the docs
:D

+++ src/generators/math/CurveGen.java
@@ -1,16 +1,30 @@
 package generators.math;
 
 public class CurveGen {
-	private final int multiplier;
-	private final int shift;
-
-	public CurveGen(final int i, final int j) {
+	private final byte multiplier;
+	private final byte shift;
+	
+	/**
+	 * This is a class that generates a "normal like" distribution
+	 * @param i
+	 * 		Multiplier (vertical stretch)
+	 * @param j
+	 * 		Shift (x-j)
+	 */
+	public CurveGen(final byte i, final byte j) {
 		this.multiplier = i;
 		this.shift = j;
 	}
-
+	
+	/**
+	 * Calculates the y coordinate based on the previously given distribution
+	 * @param input
+	 * 		The x coordinate
+	 * @return
+	 * 		The y coordinate
+	 */
 	public byte calculate(final byte input) {
-		int x = input - shift;
+		byte x = (byte) (input - shift);
 		return (byte) (multiplier * ((Math.abs(x) + 14) / (Math.pow((.1 * (x) - 1), 2) + 5)));
 	}
